Pass token as argument directly
Instead of trying to parse stdin, this will pass in the token as an
argument correctly. In terms of not being able to extract this from
outside the CodeQL CLI this should be completely safe since the args
are passed to the CLI server and are never actually passed to a
command-line application, they are only passed over stdin.

Passing the token directly is simpler and more resilient to changes in
the CodeQL CLI. However, the CodeQL CLI will now log the token in its
output since it logs the arguments that are passed to the plumbing
commands. We can remove these from the logs by redacting the token, but
it is easy to miss a code path, in which case the token would still be
logged.